Fix floating-point precision bug in transaction amounts
Transaction.Amount, its DTOs, and TransactionSummaryDto's totals were
all `double` -- binary floating-point, which can't exactly represent
most decimal fractions (0.1 + 0.2 famously comes out to
0.30000000000000004). For a ledger summing many transactions into a
balance, those tiny rounding errors compound. Switched everything to
`decimal`, the base-10 type built for financial math, with an explicit
decimal(18,2) column mapping so EF Core doesn't fall back to a
provider default. Added an EF Core migration and applied it.

Also adds a Ledger.API.Tests project (previously none existed) with a
test proving GetSummary now sums fractional amounts with exact
decimal precision, plus a second test covering the per-user ownership
filter on the same method.

+++ Ledger.API/Data/Services/TransactionsService.cs
@@ -64,8 +64,8 @@ namespace Ledger.API.Data.Services
         {
             var userTransactions = context.Transactions.Where(t => t.UserId == userId);
 
-            var totalIncome = userTransactions.Where(t => t.Type == "Income").Sum(t => (double?)t.Amount) ?? 0;
-            var totalExpenses = userTransactions.Where(t => t.Type == "Expense").Sum(t => (double?)t.Amount) ?? 0;
+            var totalIncome = userTransactions.Where(t => t.Type == "Income").Sum(t => (decimal?)t.Amount) ?? 0;
+            var totalExpenses = userTransactions.Where(t => t.Type == "Expense").Sum(t => (decimal?)t.Amount) ?? 0;
 
             return new TransactionSummaryDto
             {
